+++ plugins/dbwv/messages/de/admin.php
@@ -6,5 +6,6 @@ return [
     'motion_topics' => 'Sachtebiete',
     'con_topics' => 'Sachgebiete',
     'con_topic_add' => 'Neues Sachgebiet',
-    'list_tag' => 'Sachgebiet',
+    'list_tag' => 'Sach- / Themengebiet',
+    'filter_tag' => 'Sach- / Themengebiet',
 ];